fix(webapp): move transaction-resilience config out of db.server so wholesale db.server mocks don't break
